zmq i/o threads tuning

+++ services/lake/src/socket.rs
@@ -19,7 +19,7 @@ impl Context {
         }
     }
 
-    pub fn set_io_threads(&self, value: i32) -> Result<(), error::Error> {
+    pub fn set_io_threads(&self, value: usize) -> Result<(), error::Error> {
         if unsafe {
             zmq_sys::zmq_ctx_set(self.underlying, zmq_sys::ZMQ_IO_THREADS as _, value as i32)
         } == -1
